cognitus_django/requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
decorator==5.1.1
Django==3.2.14
djangorestframework==3.13.1
idna==3.3
ipdb==0.13.9
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
numpy==1.19.5
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
psycopg2==2.7.7
ptyprocess==0.7.0
Pygments==2.12.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.11
wcwidth==0.2.5
xlrd==1.2.0
